+++ app/src/main/java/campus/tech/kakao/map/model/datasource/LocationApi.kt
@@ -1,11 +1,6 @@
 package campus.tech.kakao.map.model.datasource
 
-import android.util.Log
-import campus.tech.kakao.map.model.Location
-import campus.tech.kakao.map.model.Location.Companion.toLocation
-import campus.tech.kakao.map.model.LocationDto
 import campus.tech.kakao.map.model.SearchFromKeywordResponse
-import campus.tech.kakao.map.model.repository.RetrofitInstance
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
