+++ app/src/main/java/campus/tech/kakao/map/model/repository/SavedLocationRepository.kt
@@ -16,7 +16,6 @@ class SavedLocationRepository(
     }
 
     suspend fun deleteSavedLocation(savedLocation: SavedLocation) {
-
         appDatabase.savedLocationDao().delete(savedLocation)
     }
